% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{clean_ws}
\alias{clean_ws}
\title{Sets up directory structure for a new Borges site
clean_ws}
\usage{
clean_ws(keep = c())
}
\arguments{
\item{pos}{string, "first" or "last"}
}
\value{
bool
}
\description{
is.empty
}
\examples{
\dontrun{
}
}

