% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dirs-files.R
\name{copy_dir}
\alias{copy_dir}
\title{Copy directories recursively, creating a new directory if not already there}
\usage{
copy_dir(from, to, overwrite = TRUE)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
copy_dir
}
\examples{
\dontrun{
}
}

