% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dirs-files.R
\name{copy_dirs}
\alias{copy_dirs}
\title{Copy directories recursively, creating a new directory if not already there}
\usage{
copy_dirs(from, to)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
copy_dirs
}
\examples{
\dontrun{
}
}

