% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{copy_files}
\alias{copy_files}
\title{Copy directories recursively, creating a new directory if not already there}
\usage{
copy_files(from, to, overwrite = TRUE)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
copy_files
}
\examples{
\dontrun{
}
}

