% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dirs-files.R
\name{list_files}
\alias{list_files}
\title{List directories in a directory}
\usage{
list_files(path = ".", recursive = TRUE, fullPath = TRUE)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
list_files
}
\examples{
\dontrun{
}
}

