% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{list_to_dataframe2}
\alias{list_to_dataframe2}
\title{List to Data Frame}
\usage{
list_to_dataframe2(l)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
list_to_dataframe2
}
\examples{
\dontrun{
}
}

