% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dirs-files.R
\name{read_sysfile}
\alias{read_sysfile}
\title{Read contents of a system file into a character string}
\usage{
read_sysfile(..., package = "utter")
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
read_sysfile
}
\examples{
\dontrun{
}
}

