% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exports.R
\name{this if null}
\alias{\%||\%}
\alias{this if null}
\title{%||%}
\usage{
x \%||\% y
}
\arguments{
\item{x}{any}

\item{y}{any}
}
\value{
y if x is undefined
}
\description{
this or that: %||%
}
\examples{
\dontrun{
}
}

